+++ CustomAI/RealCityOfficeBuildingAI.cs
@@ -3,7 +3,7 @@ namespace RealCity.CustomAI
 {
     public class RealCityOfficeBuildingAI
     {
-        public static TransferManager.TransferReason OfficeBuildingAIGetOutgoingTransferReasonPreFix()
+        public TransferManager.TransferReason GetOutgoingTransferReason()
         {
                 return TransferManager.TransferReason.None;
         }
