+++ CustomAI/RealCityPrivateBuildingAI.cs
@@ -3,11 +3,13 @@ using ColossalFramework.Math;
 using RealCity.Util;
 using RealCity.UI;
 using System;
+using Harmony;
+using System.Reflection;
 
 namespace RealCity.CustomAI
 {
-
-    public class RealCityPrivateBuildingAI
+    [HarmonyPatch]
+    public static class RealCityPrivateBuildingAI
     {
         public static ushort allBuildings = 0;
         public static uint preBuidlingId = 0;
@@ -29,6 +31,11 @@ namespace RealCity.CustomAI
 
         public static byte[] saveData = new byte[44];
 
+        public static MethodBase TargetMethod()
+        {
+            return typeof(PrivateBuildingAI).GetMethod("SimulationStepActive", BindingFlags.NonPublic | BindingFlags.Instance);
+        }
+
         public static void Load()
         {
             int i = 0;
@@ -49,7 +56,6 @@ namespace RealCity.CustomAI
             profitBuildingCountFinal = SaveAndRestore.load_ushort(ref i, saveData);
             DebugLog.LogToFileOnly("saveData in private building is " + i.ToString());
         }
-
 
         public static void save()
         {
@@ -75,81 +81,13 @@ namespace RealCity.CustomAI
             SaveAndRestore.save_ushort(ref i, profitBuildingCount, ref saveData);
             SaveAndRestore.save_ushort(ref i, profitBuildingCountFinal, ref saveData);
         }
-        //TODO: move to harmony
-        public static void PrivateBuildingAISimulationStepActivePostFix(ushort buildingID, ref Building buildingData, ref Building.Frame frameData)
+
+        public static void Postfix(ushort buildingID, ref Building buildingData, ref Building.Frame frameData)
         {
-            //TODO, use harmony to detour this.
             ProcessLandFee(buildingData, buildingID);
             LimitAndCheckBuildingMoney(buildingData, buildingID);
             ProcessBuildingDataFinal(buildingID, ref buildingData);
             ProcessAdditionProduct(buildingID, ref buildingData);
-        }
-
-        public static long GetResidentialBuildingAverageMoney(ushort buildingID, ref Building buildingData)
-        {
-            CitizenManager instance = Singleton<CitizenManager>.instance;
-            uint num = buildingData.m_citizenUnits;
-            int num2 = 0;
-            long totalMoney = 0;
-            long averageMoney = 0;
-            while (num != 0u)
-            {
-                if ((ushort)(instance.m_units.m_buffer[(int)((UIntPtr)num)].m_flags & CitizenUnit.Flags.Home) != 0)
-                {
-                    if((instance.m_units.m_buffer[(int)((UIntPtr)num)].m_citizen0 != 0) || (instance.m_units.m_buffer[(int)((UIntPtr)num)].m_citizen1 != 0) || (instance.m_units.m_buffer[(int)((UIntPtr)num)].m_citizen2 != 0) || (instance.m_units.m_buffer[(int)((UIntPtr)num)].m_citizen3 != 0) || (instance.m_units.m_buffer[(int)((UIntPtr)num)].m_citizen4 != 0))
-                    {
-                        num2++;
-                        totalMoney += (long)MainDataStore.family_money[num];
-                    }
-                }
-                num = instance.m_units.m_buffer[(int)((UIntPtr)num)].m_nextUnit;
-                if (++num2 > 524288)
-                {
-                    CODebugBase<LogChannel>.Error(LogChannel.Core, "Invalid list detected!\n" + Environment.StackTrace);
-                    break;
-                }
-            }
-
-            if (num2 != 0)
-            {
-                averageMoney = totalMoney / num2;
-            }
-
-            return averageMoney;
-        }
-
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Redundancy", "RCS1213", Justification = "Harmony patch")]
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Naming Rules", "SA1313", Justification = "Harmony patch")]
-        public static void PrivateBuildingAIGetColorPostFix(ushort buildingID, ref Building data, InfoManager.InfoMode infoMode, ref UnityEngine.Color __result)       {
-            if (infoMode == InfoManager.InfoMode.LandValue)
-            {
-                ItemClass @class = data.Info.m_class;
-                ItemClass.Service service = @class.m_service;
-                switch (service)
-                {
-                    case ItemClass.Service.Residential:
-                        long family_money = GetResidentialBuildingAverageMoney(buildingID, ref data);
-                        if (family_money < 5000)
-                            MainDataStore.building_money_threat[buildingID] = 1.0f - family_money / 10000.0f;
-                        else
-                            MainDataStore.building_money_threat[buildingID] = (15000.0f - family_money) / 20000.0f;
-
-                        if (MainDataStore.building_money_threat[buildingID] < 0.5f)
-                            __result =  UnityEngine.Color.Lerp(UnityEngine.Color.green, UnityEngine.Color.yellow, MainDataStore.building_money_threat[buildingID] * 2.0f);
-                        else
-                            __result = UnityEngine.Color.Lerp(UnityEngine.Color.yellow, UnityEngine.Color.red, (MainDataStore.building_money_threat[buildingID] - 0.5f) * 2.0f);
-                        break;
-
-                    case ItemClass.Service.Office:
-                    case ItemClass.Service.Industrial:
-                    case ItemClass.Service.Commercial:
-                        if (MainDataStore.building_money_threat[buildingID] < 0.5f)
-                            __result = UnityEngine.Color.Lerp(UnityEngine.Color.green, UnityEngine.Color.yellow, MainDataStore.building_money_threat[buildingID] * 2.0f);
-                        else
-                            __result = UnityEngine.Color.Lerp(UnityEngine.Color.yellow, UnityEngine.Color.red, (MainDataStore.building_money_threat[buildingID] - 0.5f) * 2.0f);
-                        break;
-                }
-            }
         }
 
         public static void ProcessAdditionProductIndustrialExtractorAI(ushort buildingID, ref Building buildingData)
