+++ CustomAI/RealCityResidentAI.cs
@@ -1351,25 +1351,5 @@ namespace RealCity.CustomAI
                 }
             }
         }
-
-        public static void ResidentAISimulationStepPostFix(uint citizenID, ref Citizen data)
-        {
-            //change wealth
-            BuildingManager instance = Singleton<BuildingManager>.instance;
-            ushort homeBuilding = data.m_homeBuilding;
-            uint homeId = data.GetContainingUnit(citizenID, instance.m_buildings.m_buffer[homeBuilding].m_citizenUnits, CitizenUnit.Flags.Home);
-            if (MainDataStore.family_money[homeId] > 20000)
-            {
-                data.WealthLevel = Citizen.Wealth.High;
-            }
-            else if (MainDataStore.family_money[homeId] < 5000)
-            {
-                data.WealthLevel = Citizen.Wealth.Low;
-            }
-            else
-            {
-                data.WealthLevel = Citizen.Wealth.Medium;
-            }
-        }
     }
 }
