--- CustomManager/RealCityCitizenManager.cs
@@ -1,25 +1,0 @@
-﻿using RealCity.CustomAI;
-using RealCity.Util;
-
-namespace RealCity.CustomManager
-{
-    public class RealCityCitizenManager:CitizenManager
-    {
-        public static void CitizenManagerReleaseCitizenImplementationPostFix(uint citizen)
-        {
-            MainDataStore.citizenMoney[citizen] = 0;
-            MainDataStore.isCitizenFirstMovingIn[citizen] = false;
-        }
-
-        public static void CitizenManagerReleaseUnitCitizenPostFix(uint unit)
-        {
-            MainDataStore.familyGoods[unit] = 0;
-            MainDataStore.family_money[unit] = 0;
-        }
-
-        public static void CitizenManagerReleaseCitizenInstanceImplementationPreFix(ushort instance, ref CitizenInstance data)
-        {
-            RealCityHumanAI.watingPathTime[instance] = 0;
-        }
-    }
-}
