+++ Patch/CitizenManagerReleaseCitizenInstanceImplementationPatch.cs
@@ -1,0 +1,18 @@
+﻿using Harmony;
+using System.Reflection;
+
+namespace RealCity.Patch
+{
+    [HarmonyPatch]
+    public static class CitizenManagerReleaseCitizenInstanceImplementationPatch
+    {
+        public static MethodBase TargetMethod()
+        {
+            return typeof(CitizenManager).GetMethod("ReleaseCitizenInstanceImplementation", BindingFlags.NonPublic | BindingFlags.Instance);
+        }
+        public static void Prefix(ushort instance, ref CitizenInstance data)
+        {
+            HumanAISimulationStepPatch.watingPathTime[instance] = 0;
+        }
+    }
+}
