+++ Patch/CommonBuildingAIReleaseBuildingPatch.cs
@@ -1,10 +1,17 @@
-﻿using RealCity.Util;
+﻿using Harmony;
+using RealCity.Util;
+using System.Reflection;
 
-namespace RealCity.CustomAI
+namespace RealCity.Patch
 {
-    public class RealCityCommonBuildingAI: BuildingAI
+    [HarmonyPatch]
+    public static class CommonBuildingAIReleaseBuildingPatch
     {
-        public static void CommonBuildingAIReleaseBuildingPostfix(ushort buildingID)
+        public static MethodBase TargetMethod()
+        {
+            return typeof(CommonBuildingAI).GetMethod("ReleaseBuilding");
+        }
+        public static void Postfix(ushort buildingID)
         {
             MainDataStore.building_money[buildingID] = 0;
             MainDataStore.building_buffer2[buildingID] = 0;
