+++ Patch/EconomyManagerPeekResourcePatch.cs
@@ -1,0 +1,21 @@
+﻿using Harmony;
+using System.Reflection;
+
+namespace RealCity.Patch
+{
+    [HarmonyPatch]
+    public static class EconomyManagerPeekResourcePatch
+    {
+        public static MethodBase TargetMethod()
+        {
+            return typeof(EconomyManager).GetMethod("PeekResource", BindingFlags.Public | BindingFlags.Instance);
+        }
+        public static void Prefix(EconomyManager.Resource resource, ref int amount)
+        {
+            if (resource == EconomyManager.Resource.Maintenance)
+            {
+                amount *= 100;
+            }
+        }
+    }
+}
